cast written as size_t

+++ code/logic/output.c
@@ -628,7 +628,7 @@ static int fossil_io_format_internal(
     // Step 2: if markup disabled → plain copy
     if (!apply_markup)
     {
-        size_t len = (written < (int)size - 1) ? written : (size - 1);
+        size_t len = ((size_t)written < size - 1) ? (size_t)written : (size - 1);
         memcpy(buffer, temp, len);
         buffer[len] = '\0';
         return (int)len;
